Raise error for incorrect precision (#1055)
Raise an error when the precision of a variable is changed. Fixes #956.

+++ tests/epyccel/test_epyccel_functions.py
@@ -308,8 +308,9 @@ def test_multiple_returns_f14(language):
 def test_decorator_f15(language):
     @types('bool', 'int8', 'int16', 'int32', 'int64')
     def f15(a,b,c,d,e):
+        from numpy import int64
         if a:
-            return b + c
+            return int64(b + c)
         else:
             return d + e
